+++ ReactApp/server/index.js
@@ -28,6 +28,7 @@ import {
     template,
     router,
     redux,
+    // store,
     client,
     server,
 } from '__KOOT_PROJECT_CONFIG_PATHNAME__'
@@ -70,6 +71,7 @@ if (global.kootTest || (process.env.KOOT_TEST_MODE && JSON.parse(process.env.KOO
         template,
         router,
         redux,
+        // store,
         client,
         server,
     }).catch(err => {