+++ defaults/before-build.js
@@ -1,0 +1,13 @@
+module.exports = {
+    keyFileProjectConfigTemp: '__FILE_PROJECT_CONFIG_TEMP__',
+    filenameProjectConfigTemp: '.koot.config.tmp.*.js',
+    propertiesToExtract: [
+        ['name', ''],
+        ['type', 'react'],
+        ['template', ''],
+        ['router', ''],
+        ['redux', {}],
+        ['client', {}],
+        ['server', {}]
+    ]
+}