+++ test/react-isomorphic/test.js
@@ -13,12 +13,18 @@ const projectsToUse = projects.filter(project => (
     Array.isArray(project.type) && project.type.includes('react-isomorphic')
 ))
 const { changeLocaleQueryKey } = require('../../defaults/defines')
+const removeTempProjectConfig = require('../../libs/remove-temp-project-config')
 // const sleep = require('../../utils/sleep')
 
 //
 
 global.kootTest = true
 process.env.KOOT_TEST_MODE = JSON.stringify(true)
+
+//
+
+/** @type {Boolean} 是否进行完整测试。如果为否，仅测试一次打包结果 */
+const fullTest = true
 
 //
 
@@ -68,32 +74,40 @@ const testPage = async (port) => {
     const page = await browser.newPage()
     const url = isNaN(port) ? port : `http://127.0.0.1:${port}`
 
+    // 测试: 同构结果
     {
         const res = await page.goto(url, {
             waitUntil: 'networkidle0'
         }).catch()
-        const pageTitle = await page.evaluate(() => document.querySelector('title').innerText)
-        const $app = await page.$('#app')
+        const pageContent = await page.content()
 
+        // 测试: 请求应 OK
         expect(res.ok()).toBe(true)
+
+        // 测试: 页面标题的注入应成功
+        const pageTitle = await page.evaluate(() => document.querySelector('title').innerText)
         expect(typeof pageTitle).toBe('string')
+
+        // 测试: #app 是否存在
+        const $app = await page.$('#app')
         expect(typeof $app).toBe('object')
-    }
-    {
-        await page.goto(`${url}?${changeLocaleQueryKey}=zh`, {
-            waitUntil: 'networkidle0'
-        })
-        const localeId = await page.evaluate(() => document.querySelector('meta[name="koot-locale-id"]').getAttribute('content'))
-        expect(localeId).toBe('zh')
-    }
-    {
-        await page.goto(`${url}?${changeLocaleQueryKey}=en`, {
-            waitUntil: 'networkidle0'
-        })
-        const localeId = await page.evaluate(() => document.querySelector('meta[name="koot-locale-id"]').getAttribute('content'))
-        expect(localeId).toBe('en')
+
+        // 测试: <script> 标签之间不应有 ,
+        expect(/<\/script>,<script/g.test(pageContent)).toBe(false)
     }
 
+    // 测试: 利用强制切换语种 URL 访问时，语种应正确
+    const testLocaleIdByQuery = async (localeId) => {
+        await page.goto(`${url}?${changeLocaleQueryKey}=${localeId}`, {
+            waitUntil: 'networkidle0'
+        })
+        const theLocaleId = await page.evaluate(() => document.querySelector('meta[name="koot-locale-id"]').getAttribute('content'))
+        expect(theLocaleId).toBe(localeId)
+    }
+    await testLocaleIdByQuery('zh')
+    await testLocaleIdByQuery('en')
+
+    // 测试: 到其他语种的链接
     const testLinksToOtherLang = async (urlAppend) => {
         await page.goto(`${url}${urlAppend}`, {
             waitUntil: 'networkidle0'
@@ -183,98 +197,102 @@ describe('测试: React 同构项目', async () => {
                 await testPage(port)
                 await terminate(child.pid)
             })
-            test(`[Production] 使用 koot-start (--no-build) 命令启动服务器并访问 (自定义端口号)`, async () => {
-                const port = '8316'
-                const commandName = `${commandTestBuild}-isomorphic-start-server-custom-port`
-                const command = `koot-start --no-build --port ${port}`
-                await addCommand(commandName, command, dir)
+            if (fullTest) {
+                test(`[Production] 使用 koot-start (--no-build) 命令启动服务器并访问 (自定义端口号)`, async () => {
+                    const port = '8316'
+                    const commandName = `${commandTestBuild}-isomorphic-start-server-custom-port`
+                    const command = `koot-start --no-build --port ${port}`
+                    await addCommand(commandName, command, dir)
 
-                const child = execSync(
-                    `npm run ${commandName}`,
-                    {
-                        cwd: dir,
-                    },
-                    // (err, stdout, stderr) => {
-                    //     console.log('err', err)
-                    //     console.log('stdout', stdout)
-                    //     console.log('stderr', stderr)
-                    // }
-                )
-                const errors = []
+                    const child = execSync(
+                        `npm run ${commandName}`,
+                        {
+                            cwd: dir,
+                        },
+                        // (err, stdout, stderr) => {
+                        //     console.log('err', err)
+                        //     console.log('stdout', stdout)
+                        //     console.log('stderr', stderr)
+                        // }
+                    )
+                    const errors = []
 
-                // child.stdin.pipe(process.stdin)
-                // child.stdout.pipe(process.stdout)
-                // child.stderr.pipe(process.stderr)
-                child.stderr.on('data', err => {
-                    errors.push(err)
+                    // child.stdin.pipe(process.stdin)
+                    // child.stdout.pipe(process.stdout)
+                    // child.stderr.pipe(process.stderr)
+                    child.stderr.on('data', err => {
+                        errors.push(err)
+                    })
+
+                    // console.log(111)
+                    await waitForPort(child)
+                    // console.log(222)
+
+                    // console.log({
+                    //     port,
+                    //     errors,
+                    // })
+                    expect(errors.length).toBe(0)
+
+                    await testPage(port)
+                    await terminate(child.pid)
                 })
+                test(`[Production] 使用打包后的执行文件启动服务器并访问`, async () => {
+                    const cwd = path.resolve(dir, 'dist')
+                    const child = execSync(
+                        `node ${path.resolve(cwd, 'index.js')}`,
+                        {
+                            cwd,
+                        },
+                    )
+                    const errors = []
 
-                // console.log(111)
-                await waitForPort(child)
-                // console.log(222)
+                    const port = await waitForPort(child)
+                    child.stderr.on('data', err => {
+                        errors.push(err)
+                    })
 
-                // console.log({
-                //     port,
-                //     errors,
-                // })
-                expect(errors.length).toBe(0)
+                    // console.log({
+                    //     port,
+                    //     errors,
+                    // })
+                    expect(errors.length).toBe(0)
 
-                await testPage(port)
-                await terminate(child.pid)
-            })
-            test(`[Production] 使用打包后的执行文件启动服务器并访问`, async () => {
-                const cwd = path.resolve(dir, 'dist')
-                const child = execSync(
-                    `node ${path.resolve(cwd, 'index.js')}`,
-                    {
-                        cwd,
-                    },
-                )
-                const errors = []
-
-                const port = await waitForPort(child)
-                child.stderr.on('data', err => {
-                    errors.push(err)
+                    await testPage(port)
+                    await terminate(child.pid)
                 })
+                test(`[Development] 启动开发模式并访问`, async () => {
+                    // const port = '8316'
+                    const commandName = `${commandTestBuild}-isomorphic-dev`
+                    const command = `koot-dev --no-open`
+                    await addCommand(commandName, command, dir)
 
-                // console.log({
-                //     port,
-                //     errors,
-                // })
-                expect(errors.length).toBe(0)
+                    const child = execSync(
+                        `npm run ${commandName}`,
+                        {
+                            cwd: dir,
+                        },
+                    )
+                    const errors = []
 
-                await testPage(port)
-                await terminate(child.pid)
-            })
-            test(`[Development] 启动开发模式并访问`, async () => {
-                // const port = '8316'
-                const commandName = `${commandTestBuild}-isomorphic-dev`
-                const command = `koot-dev --no-open`
-                await addCommand(commandName, command, dir)
+                    const port = await waitForPort(child, / on.*http:.*:([0-9]+)/)
+                    child.stderr.on('data', err => {
+                        errors.push(err)
+                    })
 
-                const child = execSync(
-                    `npm run ${commandName}`,
-                    {
-                        cwd: dir,
-                    },
-                )
-                const errors = []
+                    // console.log({
+                    //     port,
+                    //     errors,
+                    // })
+                    expect(errors.length).toBe(0)
 
-                const port = await waitForPort(child, / on.*http:.*:([0-9]+)/)
-                child.stderr.on('data', err => {
-                    errors.push(err)
+                    await testPage(port)
+                    await terminate(child.pid)
                 })
+            }
 
-                // console.log({
-                //     port,
-                //     errors,
-                // })
-                expect(errors.length).toBe(0)
-
-                await testPage(port)
-                await terminate(child.pid)
-            })
+            // 移除临时项目配置文件
+            await removeTempProjectConfig(dir)
         })
-
     }
 })